+++ cluster/images/docs/src/release_notes.rst
@@ -16,6 +16,7 @@ Release Notes
     been changed to ``https://sv.sv-1.svc.TARGET_CLUSTER.network.canton.global/api/v0/sv``.
   * ``sponsorApiPort`` in ``sv-values.yaml`` has been removed. The port is now included in ``sponsorApiUrl``. The default sponsor address has
     been changed to ``https://sv.sv-1.svc.TARGET_CLUSTER.network.canton.global/api/v0/sv``.
+  * The SV private and public key are now stored in k8s secrets.
 
 2023-05-21
 ----------